tubers/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.48.0
django-ckeditor==6.2.0
django-js-asset==2.0.0
djangocms-admin-style==3.0.0
idna==3.3
oauthlib==3.2.0
Pillow==9.0.0
pkg_resources==0.0.0
psycopg==3.0.8
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.8
